Go split right
