+++ Makefile
@@ -32,9 +32,9 @@ test_rebuild:
 
 refresh:
 	./artifacts/fireback/f gen gof --def modules/abac/AbacModule3.yml --relative-to . --gof-module github.com/torabian/fireback --no-cache true && \
+	./artifacts/fireback/f gen gof --def modules/fireback/FirebackModule3.yml --relative-to . --gof-module github.com/torabian/fireback --no-cache true && \
 	./artifacts/fireback/f gen gof --def modules/suggestion/SuggestionModule3.yml --relative-to . --gof-module github.com/torabian/fireback --no-cache true && \
 	./artifacts/fireback/f gen gof --def modules/payment/PaymentModule3.yml --relative-to . --gof-module github.com/torabian/fireback --no-cache true && \
-	./artifacts/fireback/f gen gof --def modules/fireback/FirebackModule3.yml --relative-to . --gof-module github.com/torabian/fireback --no-cache true && \
 	make
 
 bundle:
