+++ modules/abac/CheckPassportMethodsAction.dyno.go
@@ -401,17 +401,6 @@ func (x CheckPassportMethodsActionRequest) IsGin() bool {
 func CheckPassportMethodsActionQueryFromGin(c *gin.Context) CheckPassportMethodsActionQuery {
 	return CheckPassportMethodsActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x CheckPassportMethodsActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // CheckPassportMethodsActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the CheckPassportMethodsAction action. Developers implement
