+++ modules/abac/ClassicSignupActionCli.dyno.go
@@ -1,0 +1,17 @@
+//go:build !wasm
+
+package abac
+
+import "reflect"
+
+func (x ClassicSignupActionRequest) IsCli() bool {
+	if x.CliCtx == nil {
+		return false
+	}
+	v := reflect.ValueOf(x.CliCtx)
+	switch v.Kind() {
+	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
+		return !v.IsNil()
+	}
+	return true
+}
