+++ modules/abac/ConfirmClassicPassportTotpAction.dyno.go
@@ -405,17 +405,6 @@ func (x ConfirmClassicPassportTotpActionRequest) IsGin() bool {
 func ConfirmClassicPassportTotpActionQueryFromGin(c *gin.Context) ConfirmClassicPassportTotpActionQuery {
 	return ConfirmClassicPassportTotpActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x ConfirmClassicPassportTotpActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // ConfirmClassicPassportTotpActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the ConfirmClassicPassportTotpAction action. Developers implement
