+++ modules/abac/GsmSendSmsAction.dyno.go
@@ -391,17 +391,6 @@ func (x GsmSendSmsActionRequest) IsGin() bool {
 func GsmSendSmsActionQueryFromGin(c *gin.Context) GsmSendSmsActionQuery {
 	return GsmSendSmsActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x GsmSendSmsActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // GsmSendSmsActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the GsmSendSmsAction action. Developers implement
