+++ modules/abac/InviteToWorkspaceAction.dyno.go
@@ -320,17 +320,6 @@ func (x InviteToWorkspaceActionRequest) IsGin() bool {
 func InviteToWorkspaceActionQueryFromGin(c *gin.Context) InviteToWorkspaceActionQuery {
 	return InviteToWorkspaceActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x InviteToWorkspaceActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // InviteToWorkspaceActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the InviteToWorkspaceAction action. Developers implement
