+++ modules/abac/QueryWorkspaceTypesPubliclyAction.dyno.go
@@ -369,17 +369,6 @@ func (x QueryWorkspaceTypesPubliclyActionRequest) IsGin() bool {
 func QueryWorkspaceTypesPubliclyActionQueryFromGin(c *gin.Context) QueryWorkspaceTypesPubliclyActionQuery {
 	return QueryWorkspaceTypesPubliclyActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x QueryWorkspaceTypesPubliclyActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // QueryWorkspaceTypesPubliclyActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the QueryWorkspaceTypesPubliclyAction action. Developers implement
