+++ modules/abac/UserInvitationsAction.dyno.go
@@ -407,17 +407,6 @@ func (x UserInvitationsActionRequest) IsGin() bool {
 func UserInvitationsActionQueryFromGin(c *gin.Context) UserInvitationsActionQuery {
 	return UserInvitationsActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x UserInvitationsActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // UserInvitationsActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the UserInvitationsAction action. Developers implement
