+++ modules/abac/UserPassportsAction.dyno.go
@@ -377,17 +377,6 @@ func (x UserPassportsActionRequest) IsGin() bool {
 func UserPassportsActionQueryFromGin(c *gin.Context) UserPassportsActionQuery {
 	return UserPassportsActionQueryFromString(c.Request.URL.RawQuery)
 }
-func (x UserPassportsActionRequest) IsCli() bool {
-	if x.CliCtx == nil {
-		return false
-	}
-	v := reflect.ValueOf(x.CliCtx)
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
-		return !v.IsNil()
-	}
-	return true
-}
 
 // UserPassportsActionHttpHandler returns the HTTP method, the ServeMux pattern, and a
 // typed net/http handler for the UserPassportsAction action. Developers implement
