+++ modules/fireback/CapabilitiesTreeActionCli.dyno.go
@@ -1,0 +1,17 @@
+//go:build !wasm
+
+package fireback
+
+import "reflect"
+
+func (x CapabilitiesTreeActionRequest) IsCli() bool {
+	if x.CliCtx == nil {
+		return false
+	}
+	v := reflect.ValueOf(x.CliCtx)
+	switch v.Kind() {
+	case reflect.Ptr, reflect.Map, reflect.Slice, reflect.Interface, reflect.Func, reflect.Chan:
+		return !v.IsNil()
+	}
+	return true
+}
